Create user router
