+++ api-v2/app/api/main.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter
 
-from api.routes import auth
+from api.routes import auth, user
 
 
 api_router = APIRouter()
 api_router.include_router(auth.router, prefix="/auth", tags=["auth"])
+api_router.include_router(user.router, prefix="/user", tags=["user"])
